pyproject.toml
[tool.poetry]
name = "sprout.ai"
version = "0.1.0"
description = ""
authors = ["Samuil Plamenov <samuil.plamenov@harpercollins.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
requests-mock = "^1.9.3"
Flask = "^2.1.2"
Flask-RESTful = "^0.3.9"
python-dotenv = "^0.20.0"
coverage = "^6.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
